+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Replace `eslint-plugin-import` with `eslint-plugin-import-x` for ESLint v10 compatibility
+- Rename `import/order` rule to `import-x/order`
+
 ## v10.0.0 - 2026-03-05
 
 ### BREAKING CHANGES